moved the pagetitle component to the clanlayout.tsx
