--- frontend-next-migration/src/shared/ui/PageTitle/index.ts
@@ -1,1 +1,0 @@
-export { default as PageTitle } from './ui/PageTitle';
