--- frontend-next-migration/src/shared/ui/PageTitle/ui/PageTitle.module.scss
@@ -1,52 +1,0 @@
-.PageTitleContainer {
-    display: flex;
-    align-items:flex-end;
-    margin-bottom: 2rem;
-    padding-right: 1.5rem;
-    @media (max-width: breakpoint(lg)) {
-        display: block;
-        padding-right: unset;
-    }
-}
-
-.PageTitleLeft {
-    width: 25%;
-    margin: 0.5rem;
-    @media (max-width: breakpoint(lg)) {
-        visibility: hidden;
-        display: none;
-    }
-}
-
-.PageTitle {
-    flex: 1 1;
-    padding-left: 1.5rem;
-    max-width: 75%;
-
-    h1 {
-        color: var(--primary-color);
-        font-family: var(--heading-font);
-        font-size: var(--font-size-xxxl);
-    }
-
-    @media (max-width: breakpoint(lg)) {
-        padding-left: unset;
-        max-width: unset;
-        text-align: center;
-    }
-}
-
-.SearchContainer {
-    width: 180px;
-
-    @media (max-width: breakpoint(lg)) {
-        visibility: hidden;
-        display: none;
-    }
-
-    // properties to enhance the placeholder
-    background-color: var(--background);
-    color: var(--primary-color);
-    border-radius: var(--border-radius-lg);
-    padding: 0.1em;
-}
